+++ src/routers/keywords.py
@@ -3,7 +3,7 @@ from traceback import format_exc
 from aiogram import Router, types, filters
 from aiogram.fsm.context import FSMContext
 from aiogram.fsm.state import StatesGroup, State
-from aiogram.types import InlineKeyboardMarkup, InlineKeyboardButton
+from aiogram.types import InlineKeyboardMarkup, InlineKeyboardButton, ReplyKeyboardMarkup, KeyboardButton, ReplyKeyboardRemove
 
 import config
 from db import db
@@ -23,28 +23,41 @@ async def keywords(callback: types.CallbackQuery):
             [InlineKeyboardButton(text='Список слов 📋', callback_data='key_word_list')],
             [InlineKeyboardButton(text='◀️ Назад', callback_data='menu')]
     ])
-    await callback.message.edit_text('🔑 Меню ключевых слов, которые будут искаться в сообщениях.')
-    await callback.message.edit_reply_markup(callback.inline_message_id, markup)
+    await callback.message.answer('🔑 Меню ключевых слов, которые будут искаться в сообщениях.', reply_markup=markup)
 
 
 @router.callback_query(filters.Text('add_key_word'))
 async def add_key_word(callback: types.CallbackQuery, state: FSMContext):
-    await callback.message.answer('🔑 Укажите ключевое слово (можно несколько через запятую)')
+    markup = ReplyKeyboardMarkup(keyboard=[[KeyboardButton(text='❌ Отмена')]], resize_keyboard=True)
+    await callback.message.answer('🔑 Укажите ключевое слово (можно несколько через запятую)', reply_markup=markup)
     await state.set_state(KeyWordsState.user_input)
 
 
 @router.message(KeyWordsState.user_input)
 async def user_input(message: types.Message, state: FSMContext):
     try:
-        for key_word in message.text.lower().split(','):
-                config.key_words.append(key_word.strip())
-                db.create('key_words', key_word.strip())
-        await message.answer('✅ Успешно добавлено')
+        msg = '✅ Успешно добавлено'
+        for key_word in message.text.split('\n'):
+            if key_word == '❌ Отмена':
+                msg = 'Возвращаемся назад.'
+                break
+            if key_word in config.key_words:
+                await message.answer(f'Ключевое слово <b>{key_word}</b> уже добавлено.', parse_mode='html')
+                continue
+            config.key_words.append(key_word.strip())
+            db.create('key_words', key_word.strip())
+        await message.answer(msg, reply_markup=ReplyKeyboardRemove())
     except:
-        await message.answer('❌ Ошибка при добавлении')
+        await message.answer('❌ Ошибка при добавлении', reply_markup=ReplyKeyboardRemove())
         print(format_exc())
     finally:
         await state.clear()
+        markup = InlineKeyboardMarkup(inline_keyboard=[
+            [InlineKeyboardButton(text='Добавить слово 📥', callback_data='add_key_word'), InlineKeyboardButton(text='Удалить слово 🗑', callback_data='delete_key_word')],
+            [InlineKeyboardButton(text='Список слов 📋', callback_data='key_word_list')],
+            [InlineKeyboardButton(text='◀️ Назад', callback_data='menu')]
+        ])
+        await message.answer('🔑 Меню ключевых слов, которые будут искаться в сообщениях.', reply_markup=markup)
 
 
 @router.callback_query(filters.Text('key_word_list'))
@@ -62,29 +75,42 @@ async def key_word_list(callback: types.CallbackQuery):
 
 @router.callback_query(filters.Text('delete_key_word'))
 async def delete_key_word(callback: types.CallbackQuery, state: FSMContext):
-     msg = '🗑 Укажите номер слова, которое нужно удалить (можно несколько, через запятую)\n\n'
+    msg = '🗑 Укажите номер слова, которое нужно удалить (можно несколько, через запятую)\n\n'
 
-     for index, key_word in enumerate(config.key_words):
+    for index, key_word in enumerate(config.key_words):
         msg += f'{index+1}. {key_word}\n'
      
-     if (not len(config.key_words)):
+    if (not len(config.key_words)):
         await callback.message.answer('Нет слов для удаления.')
         return
-
-     await callback.message.answer(msg)
-     await state.set_state(KeyWordsState.delete)
+    markup = ReplyKeyboardMarkup(keyboard=[[KeyboardButton(text='❌ Отмена')]], resize_keyboard=True)
+    await callback.message.answer(msg, reply_markup=markup)
+    await state.set_state(KeyWordsState.delete)
 
 
 @router.message(KeyWordsState.delete)
 async def delete(message: types.Message, state: FSMContext):
     try:
-        values = [config.key_words[int(index)-1] for index in message.text.split(',')]
+        msg = '✅ Успешно удалено'
+        values = message.text.split('\n')
         for value in values:
+            if value.strip() == '❌ Отмена':
+                msg = 'Возвращаемся назад.'
+                break
+            if value.strip() not in config.key_words:
+                await message.answer(f'Ключевое слово <b>{value.strip()}</b> отсутствует', parse_mode='html')
+                continue
             config.key_words.remove(value)
             db.delete('key_words', 'word', value)
-        await message.answer('✅ Успешно удалено')
+        await message.answer(msg, reply_markup=ReplyKeyboardRemove())
     except:
-        await message.answer('❌ Ошибка при удалении')
+        await message.answer('❌ Ошибка при удалении', reply_markup=ReplyKeyboardRemove())
         print(format_exc())
     finally:
         await state.clear()
+        markup = InlineKeyboardMarkup(inline_keyboard=[
+            [InlineKeyboardButton(text='Добавить слово 📥', callback_data='add_key_word'), InlineKeyboardButton(text='Удалить слово 🗑', callback_data='delete_key_word')],
+            [InlineKeyboardButton(text='Список слов 📋', callback_data='key_word_list')],
+            [InlineKeyboardButton(text='◀️ Назад', callback_data='menu')]
+        ])
+        await message.answer('🔑 Меню ключевых слов, которые будут искаться в сообщениях.', reply_markup=markup)
