+++ src/routers/stopwords.py
@@ -3,7 +3,7 @@ from traceback import format_exc
 from aiogram import Router, types, filters
 from aiogram.fsm.context import FSMContext
 from aiogram.fsm.state import StatesGroup, State
-from aiogram.types import InlineKeyboardMarkup, InlineKeyboardButton
+from aiogram.types import InlineKeyboardMarkup, InlineKeyboardButton, ReplyKeyboardMarkup, ReplyKeyboardRemove, KeyboardButton
 
 import config
 from db import db
@@ -23,28 +23,41 @@ async def keywords(callback: types.CallbackQuery):
             [InlineKeyboardButton(text='Список слов 📋', callback_data='stop_word_list')],
             [InlineKeyboardButton(text='◀️ Назад', callback_data='menu')]
     ])
-    await callback.message.edit_text('🔴 Меню стов-слов, которые будут искаться в сообщениях.')
-    await callback.message.edit_reply_markup(callback.inline_message_id, markup)
+    await callback.message.answer('🔴 Меню стов-слов, которые будут искаться в сообщениях.', reply_markup=markup)
 
 
 @router.callback_query(filters.Text('add_stop_word'))
 async def add_stop_word(callback: types.CallbackQuery, state: FSMContext):
-    await callback.message.answer('🔴 Укажите стоп-слово (можно несколько через запятую)')
+    markup = ReplyKeyboardMarkup(keyboard=[[KeyboardButton(text='❌ Отмена')]], resize_keyboard=True)
+    await callback.message.answer('🔴 Укажите стоп-слово (можно несколько через запятую)', reply_markup=markup)
     await state.set_state(StopWordsState.user_input)
 
 
 @router.message(StopWordsState.user_input)
 async def user_input(message: types.Message, state: FSMContext):
     try:
-        for stop_word in message.text.lower().split(','):
-                config.stop_words.append(stop_word.strip())
-                db.create('stop_words', stop_word.strip())
-        await message.answer('✅ Успешно добавлено')
+        msg = '✅ Успешно добавлено'
+        for stop_word in message.text.split('\n'):
+            if stop_word == '❌ Отмена':
+                msg = 'Возвращаемся назад.'
+                break
+            if stop_word in config.stop_words:
+                await message.answer(f'Стоп-слово <b>{stop_word}</b> уже добавлено.', parse_mode='html')
+                continue
+            config.stop_words.append(stop_word.strip())
+            db.create('stop_words', stop_word.strip())
+        await message.answer(msg, reply_markup=ReplyKeyboardRemove())
     except:
-        await message.answer('❌ Ошибка при добавлении')
+        await message.answer('❌ Ошибка при добавлении', reply_markup=ReplyKeyboardRemove())
         print(format_exc())
     finally:
         await state.clear()
+        markup = InlineKeyboardMarkup(inline_keyboard=[
+                [InlineKeyboardButton(text='Добавить слово 📥', callback_data='add_stop_word'), InlineKeyboardButton(text='Удалить слово 🗑', callback_data='delete_stop_word')],
+                [InlineKeyboardButton(text='Список слов 📋', callback_data='stop_word_list')],
+                [InlineKeyboardButton(text='◀️ Назад', callback_data='menu')]
+        ])
+        await message.answer('🔴 Меню стов-слов, которые будут искаться в сообщениях.', reply_markup=markup)
 
 
 @router.callback_query(filters.Text('stop_word_list'))
@@ -70,21 +83,34 @@ async def delete_stop_word(callback: types.CallbackQuery, state: FSMContext):
     if (not len(config.stop_words)):
         await callback.message.answer('Нет слов для удаления.')
         return
-
-    await callback.message.answer(msg)
+    markup = ReplyKeyboardMarkup(keyboard=[[KeyboardButton(text='❌ Отмена')]], resize_keyboard=True)
+    await callback.message.answer(msg, reply_markup=markup)
     await state.set_state(StopWordsState.delete)
 
 
 @router.message(StopWordsState.delete)
 async def delete(message: types.Message, state: FSMContext):
     try:
-        values = [config.stop_words[int(index)-1] for index in message.text.split(',')]
+        msg = '✅ Успешно удалено'
+        values = message.text.split('\n')
         for value in values:
+            if value.strip() == '❌ Отмена':
+                msg = 'Возвращаемся назад.'
+                break
+            if value.strip() not in config.stop_words:
+                await message.answer(f'Стоп-слово <b>{value.strip()}</b> отсутствует', parse_mode='html')
+                continue
             config.stop_words.remove(value)
             db.delete('stop_words', 'word', value)
-        await message.answer('✅ Успешно удалено')
+        await message.answer(msg, reply_markup=ReplyKeyboardRemove())
     except:
-        await message.answer('❌ Ошибка при удалении')
+        await message.answer('❌ Ошибка при удалении', reply_markup=ReplyKeyboardRemove())
         print(format_exc())
     finally:
         await state.clear()
+        markup = InlineKeyboardMarkup(inline_keyboard=[
+                [InlineKeyboardButton(text='Добавить слово 📥', callback_data='add_stop_word'), InlineKeyboardButton(text='Удалить слово 🗑', callback_data='delete_stop_word')],
+                [InlineKeyboardButton(text='Список слов 📋', callback_data='stop_word_list')],
+                [InlineKeyboardButton(text='◀️ Назад', callback_data='menu')]
+        ])
+        await message.answer('🔴 Меню стов-слов, которые будут искаться в сообщениях.', reply_markup=markup)
